requirements.txt
opencv-python
numpy
pyvirtualcam
Pillow
